+++ src/main/java/com/prgrms/commandLineApplication/voucher/Voucher.java
@@ -3,8 +3,8 @@ package com.prgrms.commandLineApplication.voucher;
 import java.util.UUID;
 
 public interface Voucher {
-  long discount(long price);
+  double discount(double price);
   UUID getVoucherId();
   String getVoucherType();
-  long getDiscountAmount();
+  double getDiscountAmount();
 }
